refactor(code_audit): split mod.rs god-file into submodules
